22890

See merge request akita/akita-erp!106

+++ database/migrations/2019_03_05_074854_create_mst_role_auths_table.php
@@ -17,8 +17,7 @@ class CreateMstRoleAuthsTable extends Migration
             $table->increments('id');
             $table->integer('mst_role_id');
             $table->integer('mst_screen_id');
-            $table->decimal('accessible_kb',1,0)->nullable();
-//            $table->boolean('screen_disp_auth_kb')->nullable();
+            $table->integer('accessible_kb')->nullable();
             $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('modified_at')->default(DB::raw('CURRENT_TIMESTAMP'));
         });
